+++ presentation/src/main/java/com/threegap/bitnagil/presentation/recommendroutine/contract/RecommendRoutineState.kt
@@ -1,8 +1,9 @@
-package com.threegap.bitnagil.presentation.recommendroutine.model
+package com.threegap.bitnagil.presentation.recommendroutine.contract
 
 import com.threegap.bitnagil.domain.recommendroutine.model.EmotionMarbleType
 import com.threegap.bitnagil.domain.recommendroutine.model.RecommendCategory
 import com.threegap.bitnagil.domain.recommendroutine.model.RecommendLevel
+import com.threegap.bitnagil.presentation.recommendroutine.model.RecommendRoutineUiModel
 
 data class RecommendRoutineState(
     val isLoading: Boolean,
